pyproject.toml
[tool.poetry]
name = "Image2Model"
version = "0.1.0"
description = "2D image to 3D model web-service using machine learning"
authors = ["rrraumpanzer <ayylmao3713@gmail.com>","GabTenma <hoseamattews24@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
httpie = "^3.2.3"
pillow = "^11.0.0"
python-multipart = "^0.0.16"
htmlx = "^0.0.3"
pydantic = "^2.9.2"


[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
